### 2019-05-22 #1472 @whedon
[PRE REVIEW]: f90nml - A Python module for Fortran namelists
**Submitting author:** @marshallward (<a href="http://orcid.org/0000-0002-4944-7431">Marshall Ward</a>)
**Repository:** <a href="https://github.com/marshallward/f90nml" target ="_blank">https://github.com/marshallward/f90nml</a>
**Version:** v1.1
