### 2019-05-22 #1472 @danielskatz
👋 @marshallward - I think a bibliography is needed by @whedon - perhaps you can reference the Fortran standard if nothing else seems appropriate.  Once you've added a .bib file and updated the .md to use it, you can enter ```@whedon generate pdf``` here as a new comment to check the build process
