### 2019-05-23 81c8915
JOSS paper update; references
Updated the JOSS paper to include an additional use case, cleanup of
Markdown syntax, removal of an older affiliation, and integration of a
BibTeX reference document, which has also been included.
