+++ papers/joss/paper.bib
@@ -1,0 +1,13 @@
+@Book{ISO:2018:1539:1,
+    author = "{ISO}",
+    title = "{ISO\slash IEC 1539-1:2018}: Information Technology
+        -- Programming languages
+        -- Fortran
+        -- Part 1: Base language",
+    publisher = "International Organization for Standardization",
+    address = "Geneva, Switzerland",
+    pages = "630",
+    day = "28",
+    month = "November",
+    year = "2018",
+}
