### 2019-05-23 #1472 @marshallward
Thank you @danielskatz, I've updated the doc to include a reference to the ISO document.  (Not sure what else to cite here, since it's a pretty low level tool).  I will direct @whedon in the next post.

Also, I believe that @zbeekman may be a suitable referee for this paper, based on his contributions to the Fortran community.
