### 2019-05-23 25f277c
Removing escape character from JOSS paper
