+++ papers/joss/paper.bib
@@ -1,9 +1,9 @@
 @Book{ISO:2018:1539:1,
     author = "{ISO}",
-    title = "{ISO\slash IEC 1539-1:2018}: Information Technology
-        -- Programming languages
-        -- Fortran
-        -- Part 1: Base language",
+    title = "{ISO\slash IEC 1539-1:2018}: {I}nformation {T}echnology
+        -- {P}rogramming languages
+        -- {F}ortran
+        -- {P}art 1: {B}ase language",
     publisher = "International Organization for Standardization",
     address = "Geneva, Switzerland",
     pages = "630",
