### 2019-05-23 #1472 @danielskatz
Please see https://github.com/marshallward/f90nml/pull/94
