### 2019-05-23 #1472 @danielskatz
For submissions to JOSS, papers should talk a bit about the research use and users of the software - see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
