### 2019-05-23 #1472 @marshallward
Given that this is more of a Fortran tool rather than addressing a specific research activity, would you recommend submitting to a different journal?
