### 2019-05-23 #1472 @zbeekman
FWIW (maybe not much!) I can see great value in research contexts for such a tool. Two immediate use cases jump to mind (although they are sort of two sides of the same coin):

1. Research code modernization & portability
2. Mixed language programming

### 1. Modernization & Portability

Namelist files are better than old school "input deck" formatted input files, but they still are not as widely used (and easy to work with) as JSON and YAML. Research codes like NASA's [FUN3D](https://fun3d.larc.nasa.gov/) rely heavily on namelists for input files as do weather & climate codes like WRF. Increasingly these types of Engineering, Geoscience & Physics codes are needing to be integrated into workflows and analysis pipelines using other tools. As such, namelists are a poor choice of data format for both input files and outputs, as they are not easily generated or read by other tools. Increasingly research code inputs may need to be assimilated from more diverse sources, and analysis with tools like Jupyter notebooks or further processing of outputs is desired.

`f90nml` can help researchers with modernization and porability in two ways:

1. Translate namelists into more modern and widely utilized (portable) data formats. These might be research code outputs, for example.
2. Provide bulk, easy translation of namelist input files for Fortran research codes utilizing tools like [JSON-Fortran](https://github.com/jacobwilliams/json-fortran) or [petaca](https://petaca.readthedocs.io/en/master/parameter_lists.html#parameter-list-input-output-using-json), which are being increasingly adopted during code modernization and ongoing development.

### 2. Mixed Language Programming

Python is often used as glue code in scientific and engineering software. For example [Helios](https://arc.aiaa.org/doi/10.2514/6.2016-1057) integrates components written in different languages, mostly Fortran and C++, and can swap in modular near-body CFD solvers for various rotorcraft and wind energy problems, including FUN3D, mentioned above. The glue code, and main driver application is written in python, but needs to generate namelist files to help talk to the Fortran based components. A framework like `f90nml` that can read and write namelists easily from Python could be attractive as a 3rd party library to be utilized in such applications.

It is my (completely unsolicited and possibly worthless) opinion that, with suitable research, edits and references, that this f90nml joss article could be crafted into one that highlights the research uses, users and applications, and is appropriate for review. (But I haven't read the draft yet, this is based solely on my opinion and familiarity with the software.)
