### 2019-05-23 #1472 @rgieseke
Pymagicc (https://doi.org/10.21105/joss.00516) has used `f90nml` and switching to it has replaced a bunch of hacky code in a previous version, thus greatly speeding up Pymagicc development.

I'm at GitHub's satellite conference right now, where it was just announced that GitHub repos will display "Used by" other projects button.

For `f90nml` there are 45 repos and 10 packages.

https://github.com/marshallward/f90nml/network/dependents
