### 2019-05-23 #1472 @danielskatz
I have no doubts about the usefulness of this software, I'm just trying to make sure if fits JOSS - I'll confer with other editors and be back soon
