### 2019-05-23 #1472 @marshallward
Thanks all for the kind feedback.  I will modify the manuscript over the next day or so and try to consolidate some of the uses cases from these projects, including the ones linked by @rgieseke.  (I was aware of a few via Google, but I'm surprised to see much more).
