### 2019-05-23 d1c121b
JOSS documentation update
Updated manuscript for JOSS, with more references and more about
motivation for research applications.

Not yet checked for validity!
