+++ papers/joss/paper.bib
@@ -8,6 +8,119 @@
     address = "Geneva, Switzerland",
     pages = "630",
     day = "28",
-    month = "November",
+    month = {nov},
     year = "2018",
+}
+
+@article{SPECMPI2007:2010,
+    author = {Müller, Matthias and van Waveren, Matthijs and Lieberman, Ron
+        and Whitney, Brian and Saito, Hideki and Kumaran, Kalyan
+        and Baron, John and Brantley, William and Parrott, Chris
+        and Elken, Tom and Feng, Huiyu and Ponder, Carl
+    },
+    year = {2010},
+    month = {02},
+    pages = {191-205},
+    title = {SPEC MPI2007 -- An application benchmark suite for parallel
+        systems using MPI},
+    volume = {22},
+    journal = {Concurrency and Computation: Practice and Experience},
+    doi = {10.1002/cpe.1535}
+}
+
+@inproceedings{ClimateFortranDev:2014,
+ author = {M{\'e}ndez, Mariano and Tinetti, Fernando G. and Overbey, Jeffrey L.},
+ title = {Climate Models: Challenges for Fortran Development Tools},
+ booktitle = {Proceedings of the 2Nd International Workshop on Software Engineering for High Performance Computing in Computational Science and Engineering},
+ series = {SE-HPCCSE '14},
+ year = {2014},
+ isbn = {978-1-4799-7035-3},
+ location = {New Orleans, Louisiana},
+ pages = {6--12},
+ numpages = {7},
+ url = {https://doi.org/10.1109/SE-HPCCSE.2014.7},
+ doi = {10.1109/SE-HPCCSE.2014.7},
+ acmid = {2691155},
+ publisher = {IEEE Press},
+ address = {Piscataway, NJ, USA},
+}
+
+!NCAR Tech. Note NCAR/TN-556+STR
+@techreport{WRF:2019,
+    doi = {10.5065/1dfh-6p97}
+    author = {Skamarock, William C. and Joseph B. Klemp and Jimy Dudhia
+        and David O. Gill and Zhiquan Liu and Judith Berner and Wei Wang
+        and Jordan G. Powers and Michael G. Duda and Dale M. Barker
+        and Xiang-Yu Huang
+    }
+    year = 2019,
+    title = {A Description of the Advanced Research WRF Version 4},
+    pages = 145,
+    institution = {National Center for Atmospheric Research}
+}
+
+
+@techreport{MOM5:2012,
+    title = {Elements of the {Modular Ocean Model (MOM)}},
+    author = {Stephen M. Griffies},
+    institution = {NOAA/Geophysical Fluid Dynamics Laboratory},
+    pages = {xiii + 632},
+    year = 2012,
+}
+
+@article{QUANTUMESPRESSO:2009,
+    doi = {10.1088/0953-8984/21/39/395502},
+    url = {https://doi.org/10.1088%2F0953-8984%2F21%2F39%2F395502},
+    year = 2009,
+    month = {sep},
+    publisher = {{IOP} Publishing},
+    volume = {21},
+    number = {39},
+    pages = {395--502},
+    author = {Paolo Giannozzi and Stefano Baroni and Nicola Bonini
+        and Matteo Calandra and Roberto Car and Carlo Cavazzoni
+        and Davide Ceresoli and Guido L Chiarotti and Matteo Cococcioni
+        and Ismaila Dabo and Andrea Dal Corso and Stefano de Gironcoli
+        and Stefano Fabris and Guido Fratesi and Ralph Gebauer
+        and Uwe Gerstmann and Christos Gougoussis and Anton Kokalj
+        and Michele Lazzeri and Layla Martin-Samos and Nicola Marzari
[…]
