### 2019-05-23 d4c5f6e
Endline commas in JOSS .bib
