+++ papers/joss/paper.bib
@@ -25,7 +25,7 @@
         systems using MPI},
     volume = {22},
     journal = {Concurrency and Computation: Practice and Experience},
-    doi = {10.1002/cpe.1535}
+    doi = {10.1002/cpe.1535},
 }
 
 @inproceedings{ClimateFortranDev:2014,
@@ -47,16 +47,16 @@
 
 !NCAR Tech. Note NCAR/TN-556+STR
 @techreport{WRF:2019,
-    doi = {10.5065/1dfh-6p97}
+    doi = {10.5065/1dfh-6p97},
     author = {Skamarock, William C. and Joseph B. Klemp and Jimy Dudhia
         and David O. Gill and Zhiquan Liu and Judith Berner and Wei Wang
         and Jordan G. Powers and Michael G. Duda and Dale M. Barker
         and Xiang-Yu Huang
-    }
+    },
     year = 2019,
     title = {A Description of the Advanced Research WRF Version 4},
     pages = 145,
-    institution = {National Center for Atmospheric Research}
+    institution = {National Center for Atmospheric Research},
 }
 
 
