### 2019-05-23 4b8b14f
Revision of JOSS paper
