+++ papers/joss/paper.md
@@ -18,32 +18,31 @@ bibliography: paper.bib
 Summary
 =======
 
-`f90nml` is a Python module used for importing, manipulating, and writing Fortran
-namelist files [@ISO:2018:1539:1].  The primary use case for this module is to
-use the `Parser` object to read a namelist file and save its contents into a
+`f90nml` is a Python module used for importing, manipulating, and writing
+Fortran namelist files [@ISO:2018:1539:1].  The primary use case for this
+module is to read a namelist file via the `Parser` and save its contents into a
 `Namelist` data structure, which is a case-insensitive subclass of a `dict`,
 Python's intrinsic associative array.  The `Namelist` object can be read and
 modified as a standard Python `dict`, and its contents can be saved as a
 formatted namelist file.
 
-Fortran remains a dominant programming language in high-performance scientific
-computing [@SPECMPI2007:2010, @ClimateFortranDev:2014] and namelists have
-been a part of the language for decades.  Namelists were an early method of
-serializing numerical data into a human-readable format, although this has
+Fortran continues to be a dominant programming language in high-performance
+scientific computing [@SPECMPI2007:2010, @ClimateFortranDev:2014] and namelists
+have been a part of the language for decades.  Namelists were an early method
+of serializing numerical data into a human-readable format, although this has
 become less practical as data sizes have increased.  In more recent times,
-namelists have been more common used for runtime configuration [@MOM5:2012,
-@WRF:2019, @QUANTUMESPRESSO:2009, @UM:2019].  The need to manage runtime
-parameters over a large ensemble of runs and to document those parameters, can
-often be reduced in large part to a problem of parsing, modifying, and storing
-the input namelists.
+namelists have been more commonly used for runtime configuration [@MOM5:2012,
+@WRF:2019, @QUANTUMESPRESSO:2009, @UM:2019].  Much of the work associated with
+managing and documenting the runtime parameters over a large ensemble of runs
+can in part be reduced to the parsing, modifying, and storing of namelists.
 
-Python has been a dominant programming language in the sciences
+Python has been a dominant programming language in the sciences in recent years
 [@PyAstro:2019], consistent with the overall trend across programming
-[@PythonSO:2017], creating a growing need to build tools in Python which can
-manage any legacy data formats.  Given the importance of Fortran in both
+[@PythonSO:2017], which has created a growing need for tools in Python which
+can manage legacy data formats.  Given the importance of Fortran in both
 historical and modern scientific computing, the ability to accurately read and
 manipulate namelists offers the ability to both archive numerical results from
-the past, and to automate the configuration of future simulations.
+the past and to automate the configuration of future simulations.
 
 An example namelist, such as the one shown below:
 ```
@@ -67,31 +66,32 @@ nml = {
     }
 }
 ```
-
-Basic features include support for all intrinsic data types, as well as
-user-defined types and multidimensional arrays.  User-defined types are
-interpreted as hierarchical tree of `Namelist`s.  Multidimensional arrays
-are saved as nested lists of lists, with the most innermost lists corresponding
-to the first dimensional index.  This reverses the index order in Python, but
-corresponds to the usual ordering in memory.
+The module supports all intrinsic data types, as well as user-defined types and
+multidimensional arrays.  User-defined types are interpreted as hierarchical
+tree of `Namelist`s.  Multidimensional arrays are saved as nested lists of
+lists, with the most innermost lists corresponding to the first dimensional
+index.  This reverses the index order in Python, but corresponds to the usual
+ordering in memory.
 
 `f90nml` also includes a `patch` feature, which allows one to modify the values
-of an existing namelist while retaining its comments or existing formatting.
-There is some limited ability to add or remove values when patching.
+of an existing namelist while retaining its comments or existing whitespace
[…]
