+++ papers/joss/paper.bib
@@ -21,28 +21,34 @@
     year = {2010},
     month = {02},
     pages = {191-205},
-    title = {SPEC MPI2007 -- An application benchmark suite for parallel
-        systems using MPI},
+    title = {SPEC {MPI2007} -- An application benchmark suite for parallel
+        systems using {MPI}},
     volume = {22},
     journal = {Concurrency and Computation: Practice and Experience},
     doi = {10.1002/cpe.1535},
 }
 
 @inproceedings{ClimateFortranDev:2014,
- author = {M{\'e}ndez, Mariano and Tinetti, Fernando G. and Overbey, Jeffrey L.},
- title = {Climate Models: Challenges for Fortran Development Tools},
- booktitle = {Proceedings of the 2Nd International Workshop on Software Engineering for High Performance Computing in Computational Science and Engineering},
- series = {SE-HPCCSE '14},
- year = {2014},
- isbn = {978-1-4799-7035-3},
- location = {New Orleans, Louisiana},
- pages = {6--12},
- numpages = {7},
- url = {https://doi.org/10.1109/SE-HPCCSE.2014.7},
- doi = {10.1109/SE-HPCCSE.2014.7},
- acmid = {2691155},
- publisher = {IEEE Press},
- address = {Piscataway, NJ, USA},
+    author = {
+        M{\'e}ndez, Mariano and Tinetti, Fernando G.
+        and Overbey, Jeffrey L.
+    },
+    title = {Climate Models: Challenges for {Fortran} Development Tools},
+    booktitle = {
+        Proceedings of the 2nd International Workshop on Software Engineering
+        for High Performance Computing in Computational Science and Engineering
+    },
+    series = {{SE-HPCCSE} '14},
+    year = {2014},
+    isbn = {978-1-4799-7035-3},
+    location = {New Orleans, Louisiana},
+    pages = {6--12},
+    numpages = {7},
+    url = {https://doi.org/10.1109/SE-HPCCSE.2014.7},
+    doi = {10.1109/SE-HPCCSE.2014.7},
+    acmid = {2691155},
+    publisher = {IEEE Press},
+    address = {Piscataway, NJ, USA},
 }
 
 !NCAR Tech. Note NCAR/TN-556+STR
@@ -54,7 +60,8 @@
         and Xiang-Yu Huang
     },
     year = 2019,
-    title = {A Description of the Advanced Research WRF Version 4},
+    title = {A Description of the {Advanced Research WRF Version 4}},
+}
     pages = 145,
     institution = {National Center for Atmospheric Research},
 }
@@ -93,7 +100,7 @@
 }
 
 @techreport{UM:2019,
-    title = {Unified Model Documentation Paper 000},
+    title = {{Unified Model} Documentation Paper 000},
     author = {
         T. Allen and V. Bell and N. Bellouin and A. Bodas-Salcedo
         and P. Cresswell and S. Dadson and M. Dalvi and J. M. Edwards
