### 2019-05-23 #1472 @danielskatz
@marshallward - We'll go ahead and review this.  Let me know when you are satisfied with the paper, and we'll get started.

Also, if you have any suggestions for potential reviewers then please let us know. [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission, but feel free to mention other people, whether with a github username or an email address, here as well.
