### 2019-05-23 #1472 @zbeekman
I think you need to ask whedon to regenerate the pdf.

I'd be happy to be a reviewer, but I won't be able to look at it until the weekend after this coming one. I'm mid review on another JOSS paper and have a number of other obligations to attend to. Damian Rouson (a friend and colleague of mine) has reviewed for JOSS in the past, (he's on the list too, I think) and is also very involved in the Fortran community. You may consider asking him if my anticipated turn around is not fast enough. (We have a lot of overlap, so it would be a bad idea for us both to be reviewers on the same paper, IMO.)
