### 2019-05-23 1bfcdee
JOSS: Attempt to fix affiliation bug?
