### 2019-05-23 #1472 @marshallward
BTW I cannot seem to get the second affiliation working, if anyone knows how to fix...
