### 2019-05-23 01a865d
JOSS: Another affiliation fix attempt
