+++ README.rst
@@ -3,7 +3,6 @@ f90nml - A Fortran namelist parser, generator, and editor
 =========================================================
 
 A Python module and command line tool for parsing Fortran namelist files
-xxx
 
 .. image:: https://travis-ci.org/marshallward/f90nml.svg?branch=master
     :target: https://travis-ci.org/marshallward/f90nml
