### 2019-05-24 #1472 @zbeekman
I’m not against it but @rouson and I collaborate a lot so having both of us
as reviewers may be redundant or offer similar perspectives. I’ll be happy
to step aside for @rouson or review w/ him, but I wanted you
@danielskatz to be aware of this.

Happy either way so long as I can punt until after May 31.

On Thu, May 23, 2019 at 8:25 PM Daniel S. Katz <notifications@github.com>
wrote:

> @whedon <https://github.com/whedon> assign @zbeekman
> <https://github.com/zbeekman> as reviewer
>
