### 2019-05-24 #1472 @danielskatz
@tclune, @kc9qey - would either of you be willing to review this submission to JOSS?  (A JOSS review covers the short paper and the code repository, using [this set of criteria](https://joss.readthedocs.io/en/latest/review_criteria.html))
