### 2019-05-24 #1472 @tclune
Dan,

I’d be happy to.

Cheers,

- Tom
