### 2019-05-24 #1474 @danielskatz
As stated in the first comment in this issue, please carry out your review in this issue by updating the your checklist (above).

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the target repository and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

Let me know if you have any problems or questions
