### 2019-05-24 #1474 @danielskatz
Quoting the comment above:

> If you cannot edit the checklist please:
>
> 1. Make sure you're logged in to your GitHub account
> 2. Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations

I'm sure the first is ok - can you make sure you did the second?
