### 2019-05-25 #1472 @rouson
@danielskatz It sounds like you have enough reviewers.  If not, another person I would recommend is Craig Rasmussen.  I'm not sure if my tagging him here work, but here goes: @rasmussn.  My experience with python is very limited so I'm probably not the best reviewer.
