### 2019-05-31 #1474 @tclune
I had one minor issue with the installation.   3 different methods are described, but the one that that avoids the requirement for root privileges did not work for me.  I have amended a [previously closed issue](https://github.com/marshallward/f90nml/issues/49).
