### 2019-05-31 #1474 @tclune
My only other negative statement is that I could not find guidelines for community contributions.   OTOH, this is a vanilla GitHub project that should easily work with standard approaches of fork+PR or opening issues.
