### 2019-05-31 #1474 @marshallward
Thanks very much for the feedback.  On the three points:

* For installation, I'm not able to reproduce the problem you're seeing, including on my work machine where I do not have root access.  But we can work through that in marshallward/f90nml#49.  (Or perhaps I just ought to remove the explicit `setup.py` install?)

* I'll prepare a Contributing guidelines file.  (Patches are so infrequent that I haven't had much need for one.)

* I'll review the API, but could you mention any examples of missing items?  I should say that I only consider the `Namelist` and `Parser` classes, along with their public functions, as part of the public API.  But if there's an expectation to fully document all functions and to integrate them into the documentation, then I'm happy to do that.
