### 2019-05-31 #1474 @tclune
@marshallward Please disregard my API comment for now.   I did not see the Makefile and was just staring at the raw files in the ./doc directory.    I'm building the documentation and will update my comment.
