### 2019-05-31 #1474 @tclune
I retract my comment about the API.    Documentation is complete and nicely done.   And I thank the author for getting me to actually look at sphinx.   I'd heard of it, but never (knowingly) used it before.
