### 2019-06-03 marshallward/f90nml#95 @zbeekman
Community guidlines
The [JOSS review] criteria state:
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

1. Please add a `CONTRIBUTING.md` or appropriate section to the README and/or utilize github issue and PR templates to document how to contribute to the software
2. In `CONTRIBUTING.md` or README, or in an issue template, please provide some minimal guidelines on reporting problems and issues
3. In `CONTRIBUTING.md` or the README describe how to seek support. It may be as simple as "file an issue or PR describing the problem".

[JOSS review]: https://github.com/openjournals/joss-reviews/issues/1474
