### 2019-06-03 marshallward/f90nml#95 @marshallward
Thanks, will finish this off tonight.
