### 2019-06-03 marshallward/f90nml#96 @zbeekman
Add a dependency on the yaml module
Trying to use `f90nml` after installing on macOS via `pip3 install f90nml` with the `-f yaml` flag results in:

```
f90nml: error: YAML module could not be found.
```

I think you should just add an explicit dependency on `pyyaml`. Or, better yet, if there's a way to add a recommended (installed by default) dependency on `pyyaml` that someone could opt out of. (I'm not familiar enough with python packaging to know for certain if this is possible one way or the other.)
