### 2019-06-03 marshallward/f90nml#96 @zbeekman
Relates to https://github.com/openjournals/joss-reviews/issues/1474
