### 2019-06-03 marshallward/f90nml#96 @zbeekman
You could always just provide a `requirements.txt` at the top level that provides it.
