### 2019-06-03 marshallward/f90nml#97 @zbeekman
Documentation for --output is wrong
The group selection documentation is repeated for the `--output` flag documentation when `--help` is passed.

```
$  f90nml --help
usage: f90nml [-h] [--version] [--group GROUP] [--variable VARIABLE] [--patch]
              [--format FORMAT] [--output OUTPUT]
              [input] [output]

...

  --group GROUP, -g GROUP
                        specify namelist group to modify. When absent, the
                        first group is used

...

  --output OUTPUT, -o OUTPUT
                        specify namelist group to modify. When absent, the
                        first group is used
```

Relates to https://github.com/openjournals/joss-reviews/issues/1474
