### 2019-06-03 #1474 @zbeekman
@danielskatz:

- I'm waiting to check one box for https://github.com/marshallward/f90nml/issues/95
- I think https://github.com/marshallward/f90nml/issues/96 relates to installation and functionality. It should at least be documentented as a dependency, and/or a more graceful failure should occur when the yaml module is missing
- https://github.com/marshallward/f90nml/issues/97 is a pretty minor pedantic annoyance

Once community/contributing is addressed, I'm happy to check the last checkbox. IMO, the last two issues are not show-stoppers, but I would categorize them as minor changes requested.

I could comment on a number of the boxes that I checked during the review process to document my justification and thought process, if that would be useful. @danielskatz please let me know if you think I should do this and whether or not there are any additional tasks for me, beyond checking off the contributing/community box once upstream provides adequate documentation.

@marshallward Nice work!
