### 2019-06-04 #1474 @danielskatz
@zbeekman:

> I could comment on a number of the boxes that I checked during the review process to document my justification and thought process, if that would be useful.

No, you don't need to do this.
