### 2019-06-04 1dbc7f1
PEP8 cleanup
