### 2019-06-04 marshallward/f90nml#95 @marshallward
I've attached a contributing readme: 5a13db9eb6715217faee91ffdb3de593ff8d0c1d

First one I've written, but hopefully it hits all the key points.
