### 2019-06-04 2942d93
CLI: Removal of -o output flag
The output flag (-o) has been removed from the CLI, since the output
file can be set as the second positional argument.
