### 2019-06-04 marshallward/f90nml#97 @marshallward
Thanks, I am not even sure why there is an output flag in this case.  This has been removed from the source and the docs, and now relies on the positional argument.  (2942d93fc0a1a8487b1c5dd8feb71e4b82dfdf11)
