### 2019-06-04 marshallward/f90nml#97 @zbeekman
I was very confused for a brief moment. I thought that by

> This has been removed from the source and the docs, and now relies on the positional argument.

you were referring to an extant commit on master, not a new one you just pushed. 😅

I'll test then close.
