### 2019-06-04 marshallward/f90nml#97 @marshallward
Sorry!  Badly worded reply :)  Thanks for catching that error, btw.
