### 2019-06-04 marshallward/f90nml#96 @zbeekman
> I'd prefer to not add it as a dependency, since there is a design goal to only use standard library dependencies (hence no `requirements.txt` file). So the error is sort of intentional.
>
> It's not a lot of trouble to optionally install it though, e.g. `requests` has a `requests[security]` install. I could similarly do a `f90nml[yaml]` install.

All of that sounds good. Basically, it would be nice to have better documentation when:

1. You request yaml output but the module isn't present, and
2. You won't be able to use the yaml capabilities w/o installing the module

Documentation & error handling would be sufficient, IMO, to resolve this.
