### 2019-06-05 e58beed
CONTRIBUTING typo fix
