### 2019-06-05 744b350
Incorporating contributing doc to sphinx docs
The contributing document was moved into the Sphinx docs, and will be
built as part of the f90nml docs.

The README now links to this file, though it may not quite work in all
cases yet.
