### 2019-06-05 6fa457e
Update URL of contributing
