### 2019-06-05 marshallward/f90nml#95 @marshallward
I've incorporated the contributing guidelines into the Sphinx docs, and added a link into the README.
