### 2019-06-05 #1474 @zbeekman
@danielskatz My review is finished.

I can confirm that the paper and the source code meet the review criteria determined by JOSS, and enthusiastically recommend this paper/package for publication. `f90nml` fills an important void for enabling mixed language Python/Fortran programming, more easily designing and executing simulation and analysis workflow pipelines, and easing legacy code modernization efforts.

My only remaining advice to the author is that the reliance of the package on the `pyyaml` module (marshallward/f90nml#96) could be better, or more prominently documented, and an error message pointing the user in the right direction if they try to produce yaml output without would be helpful. In addition, the link to the documentation in the README is well placed near the top of the document, however, creating a documentation header, even if only followed by a one-sentence paragraph, would make it more prominent.

Otherwise, the package is very well written, performs testing and installation in a very straightforward, pythonic, and canonical way. It includes an extensive automatic test suite with great code coverage, and thorough documentation.

If I can be of any additional assistance, please do not hesitate to ask.

Thanks!

- Izaak "Zaak" Beekman
