### 2019-06-05 marshallward/f90nml#96 @marshallward
Reminding myself to include a message to install PyYAML in the error, as recommended by @zbeekman
