### 2019-06-05 #1474 @danielskatz
👋 @marshallward - please take

> My only remaining advice to the author is that the reliance of the package on the pyyaml module (marshallward/f90nml#96) could be better, or more prominently documented, and an error message pointing the user in the right direction if they try to produce yaml output without would be helpful. In addition, the link to the documentation in the README is well placed near the top of the document, however, creating a documentation header, even if only followed by a one-sentence paragraph, would make it more prominent.

as advice - if you want to make changes based on this, please do.
