### 2019-06-05 #1474 @marshallward
Thanks @danielskatz, I am currently working through this and other issues with @zbeekman and @tclune in the GitHub issues cited above, it's been valuable feedback.
