### 2019-06-05 marshallward/f90nml#96 @rgieseke
Not sure if that's necessary here, but this is how Pandas handles its conditional imports and warnings:

https://github.com/pandas-dev/pandas/blob/master/pandas/io/html.py#L29

https://github.com/pandas-dev/pandas/blob/master/pandas/io/html.py#L833

Maybe something simpler also works for `f90nml`.
