### 2019-06-05 marshallward/f90nml#96 @marshallward
Thanks @rgieseke, I'd say that I'm currently doing something similar (`try`-`except ImportError`).  Interestingly, there is no reference to those libs as optional dependencies in Pandas' `setup.py`.

I'm currently trying out an optional dependency at the moment, `pip install -e . [yaml]`, along with a message explaining this in the command line.  But maybe I'm just overthinking the whole thing and should just make it a dependency :).
