### 2019-06-05 marshallward/f90nml#96 @rgieseke
Yeah, i think the Pandas workflow is that you just have to install them yourself. The docs list them as optional dependencies: https://pandas.pydata.org/pandas-docs/stable/install.html

Not having them as a dependency would save some install time (which can add up in CI) and most likely not all people need the yaml support.

So maybe the optional [yaml] install is a good compromise.
