### 2019-06-05 2421f82
Flipping flags and input in pip command
