### 2019-06-05 2038824
README: Documentation highlight
Link to the full documentation has been made more prominent.  Thanks to
Izaak Beekman for the suggestion.
