### 2019-06-10 #1474 @marshallward
Yes, only the YAML issue left at this point, which is done but not yet documented or uploaded to PyPI.  Also waiting on some feedback on the README changes but that is not essential.

Was quite literally working on it when you posted, so should hopefully be done soon (excepting user feedback).
