### 2019-06-10 0be5653
Installation update; YAML install support
YAML installation instructions were included in the README.

NOTE: YAML support was added in a prior commit, but was not documented.
