### 2019-06-10 5944397
YAML output fix
Default flow style was disabled for YAML, and the missing library test
message was updated.
