### 2019-06-10 marshallward/f90nml#96 @marshallward
I've pushed some changes which add a `f90nml[yaml]` install, additional instructions for installing the YAML support, and added this information to the error message when YAML is unavailable.

Assuming that all looks good, I'll probably close this issue.
