### 2019-06-10 b27c8f6
Adding 3.7 Travis test support
