+++ .travis.yml
@@ -6,6 +6,7 @@ python:
     - "3.4"
     - "3.5"
     - "3.6"
+    - "3.7"
     - "pypy"
     - "pypy3"
 install:
