### 2019-06-10 b2aaf8e
Removing Python 3.7, seems broken...
