### 2019-06-10 marshallward/f90nml#96 @zbeekman
It looks good, but, for me, adding `-e` causes pip installation to fail. I can do `--user` without `-e` and things seem fine. That is independent of with or without yaml.
