### 2019-06-10 marshallward/f90nml#96 @marshallward
What is the error you're seeing?  I am still not 100% clear on when the `-e` flag is appropriate.
