### 2019-06-10 marshallward/f90nml#96 @zbeekman
It looks like you need to pass a path to the `-e` flag.

```
$ pip3 install -e --user .[yaml]
--user should either be a path to a local project or a VCS url beginning with svn+, git+, hg+, or bzr+
```

Adding `.` after the `-e` flag seems to have resolved it.

From the help:
```
  -e, --editable <path/url>   Install a project in editable mode (i.e. setuptools "develop mode") from a local project path or a VCS url.
```

I don't think it's required at all for users. I would get rid of it. (But I might be wrong.)
