### 2019-06-10 marshallward/f90nml#96 @zbeekman
Interesting, if I do:

```
$ pip3 install -e .[yaml] --user
Obtaining file:///Users/ibeekman/Sandbox/f90nml
Requirement already satisfied: PyYAML in /usr/local/lib/python3.7/site-packages (from f90nml==1.1.1) (5.1)
Installing collected packages: f90nml
  Running setup.py develop for f90nml
Successfully installed f90nml
```

That seems to work fine. But the output is different than if I do:
```
$ Processing /Users/ibeekman/Sandbox/f90nml
Requirement already satisfied: PyYAML in /Users/ibeekman/Library/Python/3.7/lib/python/site-packages (from f90nml==1.1.1) (5.1.1)
Building wheels for collected packages: f90nml
  Building wheel for f90nml (setup.py) ... done
  Stored in directory: /private/var/folders/d0/9mds_bmn77j80rnyw0btcz0r0000gq/T/tmp/tmp/tmp/tmp/tmp/tmp/tmp/pip-ephem-wheel-cache-pp5r2_vn/wheels/a0/02/5d/8edb3e17962104c646d68b7e5ad387cde9e11efb61399a44c1
Successfully built f90nml
Installing collected packages: f90nml
Successfully installed f90nml-1.1.1
```

I think users would want the wheel to be built and should use the version without `-e`. (I'm no python expert, though...)
