### 2019-06-10 marshallward/f90nml#96 @marshallward
I think you've worked it out, I mixed up the `--user` and `-e` flags.  It's not even possible to use `--user` in virtualenv, so I think you've tested it in the correct way :).  (Certainly better than I did!)

I think `pip install --user -e .` is probably the way to go here, I'll test it out and make sure it's all working this time.  I'll swap those and update it ASAP.
