### 2019-06-10 11224c0
README install fix
Fixed the instructions for pip install commands using the `--user`
flags.  Thanks to Izaak Beekman for catching these.
