### 2019-06-10 marshallward/f90nml#96 @zbeekman
I think you should drop the `-e` flag, but it should work with or without it. With `-e` it won't build and install a wheel, so it may be slightly slower...
