### 2019-06-12 6587b4d
README: Removal of -e flag
