### 2019-06-12 marshallward/f90nml#96 @marshallward
Agreed, just had a bit of a test and `-e` does not actually install the package anyway; just create a meta-package that links to your directory, which is probably not the expected behavior.

I've updated the README and removed the flags, and will do a PyPI update now.
