### 2019-06-12 e98c063
Version and changelog update
