+++ f90nml/__init__.py
@@ -6,7 +6,7 @@
 from f90nml.namelist import Namelist
 from f90nml.parser import Parser
 
-__version__ = '1.1.1'
+__version__ = '1.1.2'
 
 
 def read(nml_path):
