### 2019-06-13 marshallward/f90nml#96 @marshallward
Sounds like this may all be sorted out, so I will close this issue for now.
