### 2019-06-13 #1474 @marshallward
Barring any exception from the reviewers, I think I'd say that the various issues raised have been resolved.
