### 2019-06-13 #1474 @danielskatz
@marshallward - Can you make a deposit of the current repo into an archive (e.g. zenodo via https://guides.github.com/activities/citable-code/) and tell me the DOI?  If you also need to update the version number, let me know that as well.
