### 2019-06-14 #1474 @marshallward
@danielskatz the updated DOI via Zenodo is here:  https://zenodo.org/record/3245482

    10.5281/zenodo.3245482

I've done a version update to 1.1.2 which includes the reviewer-recommended changes, feel free to update it (I can't see a way to do it myself, but happy to do so if it's possible).
