### 2019-06-14 #1474 @danielskatz
I currently got a 404 back from https://www.theoj.org/joss-papers/joss.01474/10.21105.joss.01474.pdf
This usually works after a few minutes, but I'm just getting on a plane, so @arfon, if you see this working, please feel free to close this issue for the final publication step
