### 2019-06-14 #1474 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
