### 2019-06-14 #1474 @marshallward
I'm still getting a 404, but may be solved with patience.

Also, the recommended citation also seems to be missing my first initials, but perhaps this is standard?

Anyway, don't miss your flight over these details :)
