### 2019-06-14 #1474 @danielskatz
Yes, the 404 is a caching issue. Once it fails, you need to wait a bit for it to resolve. For me, it failed on airport Wi-Fi but worked a few minutes later via phone network
