### 2019-06-14 #1474 @marshallward
Must be something like that, I can see the file directly here:

    https://github.com/openjournals/joss-papers/blob/master/joss.01474/10.21105.joss.01474.pdf

As for the initials, I did notice that my `codemeta.json` file does not include my middle initial, which means it must have pulled the info from my GitHub account or git configs. I think I used one of the Ruby tools to generate that.  I can fix it, though not sure how post-publish edits are handled here.

Though having said all that, the recommended citation also lacks my first initial, so still might be something that needs to be fixed.
