### 2019-06-14 #1474 @marshallward
I've just looked at a few others and the "Cite as:" also lacks first initials, so perhaps this is something to be handled outside of this issue :).
