### 2019-06-14 #1474 @marshallward
Also, for what it's worth, I found a typo in the paper, p2 L14 itepreted -> interpreted, and have edited my `paper.md`, but no worries if it's too late to resolve it.
