### 2019-06-14 #1474 @arfon
> I've just looked at a few others and their "Cite as:" also lacks first initials, so perhaps this is something to be handled outside of this issue :).

Yeah, this isn't unique to this paper, this is currently how we create the recommended citations for single-author papers.

I've opened a bug report [here](https://github.com/openjournals/whedon/issues/50) but I don't expect a quick fix sorry.

> Also, for what it's worth, I found a typo in the paper, p2 L14 itepreted -> interpreted, and have edited my `paper.md`, but no worries if it's too late to resolve it.

I can fix this.
