### 2019-06-14 #1474 @arfon
> Also, for what it's worth, I found a typo in the paper, p2 L14 itepreted -> interpreted, and have edited my `paper.md`, but no worries if it's too late to resolve it.

OK, the PDF is updated in https://github.com/openjournals/joss-papers/commit/afc8a4414d32388103f16c34c9e1fbe04d8c9ffc - although it might take a while to update because of the caching @danielskatz mentioned earlier.
