### 2019-06-14 #1474 @marshallward
URLs are working for me, including the typo fix.  Thanks @arfon and all others for their work!
